+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.3.3 - 
 * Add `ServerHost` parameter to `Connect-MIAServer`
+* Add support for -SkipCertificateCheck
 ## 0.3.2 - June 2022
 * Refactor calls to Invoke-RestMethod to Invoke-MIARequest
 * Consolidate .Format.ps1xml file